+++ src/main/java/edu/jhuapl/trinity/utils/metric/HaversineMetric.java
@@ -29,6 +29,6 @@ public final class HaversineMetric extends Metric {
         final double sinLat = Math.sin(0.5 * (x[0] - y[0]));
         final double sinLong = Math.sin(0.5 * (x[1] - y[1]));
         final double result = Math.sqrt(sinLat * sinLat + Math.cos(x[0]) * Math.cos(y[0]) * sinLong * sinLong);
-        return (double) (2 * Math.asin(result));
+        return 2 * Math.asin(result);
     }
 }
